Selecting and counting users from the User-Group table needs DISTINCT to avoid duplicates.
In particular when using the catch-all group ("Default Group" setting)
the queries which count and select users from the User-Group table need
the "DISTINCT" option: the catch-all group is replaced by a '%'
wild-card in the query. As users may belong by design to more than one
group counting and selecting users comes out wrong. Even worse: the many
duplicates interfere with the paging logic of the user admin-settings.

Signed-off-by: Claus-Justus Heine <himself@claus-justus-heine.de>